Setting up level 4.

+++ levels/4/ctf-halt.sh
@@ -8,11 +8,18 @@ set -o errexit
 
 DIR="$( cd "$( dirname "$0" )" && pwd)"
 PIDFILE="${DIR}/pidfile"
+CASPER_PIDFILE="${DIR}/pidfile-casper"
 
 if [ -f "${PIDFILE}" ]; then
-  PID=`cat "${PIDFILE}"`
-  rm -f "${PIDFILE}"
-  kill -9 ${PID} || true
+  if PID=`cat "${CASPER_PIDFILE}"`; then
+    rm -f "${CASPER_PIDFILE}"
+    kill -9 ${PID} || true
+  fi
+
+  if PID=`cat "${PIDFILE}"`; then
+    rm -f "${PIDFILE}"
+    kill -9 ${PID} || true
+  fi
   echo "Level is now halted."
 else
   echo "Level is already halted."
